+++ mycodo/inputs/bme280_circuitpython.py
@@ -2,7 +2,6 @@
 import copy
 
 from mycodo.inputs.base_input import AbstractInput
-from mycodo.inputs.sensorutils import calculate_altitude
 from mycodo.inputs.sensorutils import calculate_dewpoint
 from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
 from mycodo.inputs.sensorutils import convert_from_x_to_y_unit
@@ -37,10 +36,10 @@ measurements_dict = {
 
 # Input information
 INPUT_INFORMATION = {
-    'input_name_unique': 'BME280',
+    'input_name_unique': 'BME280_CP',
     'input_manufacturer': 'BOSCH',
     'input_name': 'BME280',
-    'input_library': 'Adafruit_BME280',
+    'input_library': 'Adafruit_CircuitPython_BME280',
     'measurements_name': 'Pressure/Humidity/Temperature',
     'measurements_dict': measurements_dict,
     'url_manufacturer': 'https://www.bosch-sensortec.com/bst/products/all_products/bme280',
@@ -59,8 +58,9 @@ INPUT_INFORMATION = {
     'options_disabled': ['interface'],
 
     'dependencies_module': [
-        ('pip-pypi', 'Adafruit_GPIO', 'Adafruit_GPIO'),
-        ('pip-pypi', 'Adafruit_BME280', 'git+https://github.com/adafruit/Adafruit_Python_BME280.git')
+        ('pip-pypi', 'usb.core', 'pyusb==1.1.1'),
+        ('pip-pypi', 'adafruit_extended_bus', 'Adafruit_Extended_Bus'),
+        ('pip-pypi', 'adafruit_bme280', 'Adafruit_CircuitPython_BME280')
     ],
 
     'interfaces': ['I2C'],
@@ -70,11 +70,8 @@ INPUT_INFORMATION = {
 
 
 class InputModule(AbstractInput):
-    """
-    A sensor support class that measures the BME280's humidity, temperature,
-    and pressure, them calculates the altitude and dew point
+    """ A sensor support class that measures the BME280 """
 
-    """
     def __init__(self, input_dev, testing=False):
         super(InputModule, self).__init__(input_dev, testing=testing, name=__name__)
 
@@ -84,14 +81,18 @@ class InputModule(AbstractInput):
             self.initialize_input()
 
     def initialize_input(self):
-        from Adafruit_BME280 import BME280
+        import adafruit_bme280
+        from adafruit_extended_bus import ExtendedI2C
 
-        self.sensor = BME280(
-            address=int(str(self.input_dev.i2c_location), 16),
-            busnum=self.input_dev.i2c_bus)
+        try:
+            self.sensor = adafruit_bme280.Adafruit_BME280_I2C(
+                ExtendedI2C(self.input_dev.i2c_bus),
+                address=int(str(self.input_dev.i2c_location), 16))
+        except:
+            self.logger.exception("Setting up sensor")
 
     def get_measurement(self):
-        """ Get measurements and store in the database """
+        """ Gets the measurements """
         if not self.sensor:
             self.logger.error("Input not set up")
             return
@@ -99,21 +100,21 @@ class InputModule(AbstractInput):
         self.return_dict = copy.deepcopy(measurements_dict)
 
         if self.is_enabled(0):
-            self.value_set(0, self.sensor.read_temperature())
+            self.value_set(0, self.sensor.temperature)
 
         if self.is_enabled(1):
-            self.value_set(1, self.sensor.read_humidity())
+            self.value_set(1, self.sensor.relative_humidity)
 
         if self.is_enabled(2):
-            self.value_set(2, convert_from_x_to_y_unit('hPa', 'Pa', self.sensor.read_pressure()))
+            self.value_set(2, convert_from_x_to_y_unit('hPa', 'Pa', self.sensor.pressure))
 
-        if self.is_enabled(3) and self.is_enabled(0) and self.is_enabled(1):
+        if self.is_enabled(0) and self.is_enabled(1) and self.is_enabled(3):
             self.value_set(3, calculate_dewpoint(self.value_get(0), self.value_get(1)))
 
-        if self.is_enabled(4) and self.is_enabled(2):
-            self.value_set(4, calculate_altitude(self.value_get(2)))
+        if self.is_enabled(4):
+            self.value_set(4, self.sensor.altitude)
 
-        if self.is_enabled(5) and self.is_enabled(0) and self.is_enabled(1):
+        if self.is_enabled(0) and self.is_enabled(1) and self.is_enabled(5):
             self.value_set(5, calculate_vapor_pressure_deficit(self.value_get(0), self.value_get(1)))
 
         return self.return_dict